requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
joblib==1.5.1
nltk==3.9.1
oauthlib==3.3.1
openai==1.98.0
packaging==25.0
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pytesseract==0.3.13
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
textblob==0.19.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
